+++ run_test.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+cargo build && cargo test && cargo run -- --level exhaustive --no-test && cargo run -- --level medium